+++ packages/expo-updates/cli/__tests__/assetsVerify-test.ts
@@ -7,6 +7,7 @@ import {
   getExportedMetadataHashSet,
   getFullAssetDumpAsync,
   getFullAssetDumpHashSet,
+  getMissingAssetsAsync,
 } from '../assetsVerifyAsync';
 import { BuildManifest, ExportedMetadata, FullAssetDump } from '../assetsVerifyTypes';
 
@@ -66,6 +67,27 @@ const assetMap: FullAssetDump = new Map(
       type: 'jpg',
       fileHashes: ['a7643581b54e5c60b852c5d827e629dc'],
     },
+    '3fe96d5975601486c98c48b5dec0f777': {
+      __packager_asset: true,
+      fileSystemLocation: '/Users/dlowder/iosProjects/UpdatesAPIDemo/assets/icons',
+      httpServerLocation: '/assets/assets/icons',
+      width: 24,
+      height: 24,
+      scales: [1, 2, 3],
+      files: [
+        '/Users/dlowder/iosProjects/UpdatesAPIDemo/assets/icons/bell.png',
+        '/Users/dlowder/iosProjects/UpdatesAPIDemo/assets/icons/bell@2x.png',
+        '/Users/dlowder/iosProjects/UpdatesAPIDemo/assets/icons/bell@3x.png',
+      ],
+      hash: '3fe96d5975601486c98c48b5dec0f777',
+      name: 'bell',
+      type: 'png',
+      fileHashes: [
+        '0d38747965d80a1a453de5a07b05de55',
+        '5c37045d0f36217bdee17959fc10986a',
+        '55761b0bdbfd9d4c94c1b1f690954e37',
+      ],
+    },
   })
 );
 const buildManifest: BuildManifest = {
@@ -118,7 +140,7 @@ describe(getFullAssetDumpAsync, () => {
 describe(getFullAssetDumpHashSet, () => {
   it('Converts full asset map to set', () => {
     const result = getFullAssetDumpHashSet(assetMap);
-    expect(result.size).toEqual(4);
+    expect(result.fullAssetHashSet.size).toEqual(7); // Set should contain all hashes for an image with multiple scales
   });
 });
 describe(getBuildManifestAsync, () => {
@@ -201,5 +223,85 @@ describe(getExportedMetadataHashSet, () => {
       },
     };
     expect(() => getExportedMetadataHashSet(metadataWithoutAndroid, 'android')).toThrow();
+  });
+});
+describe(getMissingAssetsAsync, () => {
+  beforeEach(() => {
+    vol.reset();
+  });
+  it('Detects missing assets', async () => {
+    vol.fromJSON(
+      {
+        'dist/assetmap.json': JSON.stringify(Object.fromEntries(assetMap)),
+        'dist/metadata.json': JSON.stringify(metadata),
+        'android/app/build/app.manifest': JSON.stringify(buildManifest),
+      },
+      projectRoot
+    );
+    const result = await getMissingAssetsAsync(
+      `${projectRoot}/android/app/build/app.manifest`,
+      `${projectRoot}/dist/metadata.json`,
+      `${projectRoot}/dist/assetmap.json`,
+      'android'
+    );
+    expect(result.length).toBe(3);
+    expect(result[1].hash).toEqual('5c37045d0f36217bdee17959fc10986a');
+    expect(result[1].path).toEqual(
+      '/Users/dlowder/iosProjects/UpdatesAPIDemo/assets/icons/bell@2x.png'
+    );
+  });
+  it('Detects no missing assets', async () => {
+    // Add the missing assets to the build manifest
+    const expandedBuildManifest = {
+      ...buildManifest,
+      assets: [
+        ...buildManifest.assets,
+        {
+          name: 'bell',
+          type: 'png',
+          scale: 1,
+          packagerHash: '0d38747965d80a1a453de5a07b05de55',
+          subdirectory: '/assets/assets/icons',
+          scales: [1, 2, 3],
+          resourcesFilename: 'assets_icons_bell',
+          resourcesFolder: 'drawable',
+        },
+        {
+          name: 'bell',
+          type: 'png',
+          scale: 2,
+          packagerHash: '5c37045d0f36217bdee17959fc10986a',
+          subdirectory: '/assets/assets/icons',
+          scales: [1, 2, 3],
+          resourcesFilename: 'assets_icons_bell',
+          resourcesFolder: 'drawable',
+        },
+        {
+          name: 'bell',
+          type: 'png',
+          scale: 3,
+          packagerHash: '55761b0bdbfd9d4c94c1b1f690954e37',
+          subdirectory: '/assets/assets/icons',
+          scales: [1, 2, 3],
+          resourcesFilename: 'assets_icons_bell',
+          resourcesFolder: 'drawable',
+        },
+      ],
+    };
+    vol.fromJSON(
+      {
+        'dist/assetmap.json': JSON.stringify(Object.fromEntries(assetMap)),
+        'dist/metadata.json': JSON.stringify(metadata),
+        'android/app/build/app.manifest': JSON.stringify(expandedBuildManifest),
+      },
+      projectRoot
+    );
+    const result = await getMissingAssetsAsync(
+      `${projectRoot}/android/app/build/app.manifest`,
+      `${projectRoot}/dist/metadata.json`,
+      `${projectRoot}/dist/assetmap.json`,
+      'android'
+    );
+    expect(result.length).toBe(0);
   });
 });